requirements.txt
djangorestframework
aniso8601==7.0.0
certifi==2019.6.16
Django==2.2.4
django-cors-headers==3.0.0
django-filter==2.2.0
django-templated-mail==1.1.1
djoser==2.0.3
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
